requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==4.0.0
boto3==1.26.163
botocore==1.29.165
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.3.0
PyMySQL==1.1.0
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.12
pydantic_core==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
sqladmin==0.12.0
SQLAlchemy==2.0.21
sqlalchemy-json==0.5.0
starlette==0.27.0
typing_extensions==4.8.0
ulid-py==1.1.0
urllib3==1.26.16
uvicorn==0.22.0
WTForms==3.0.1
zipp==3.17.0